refactor prices module
