+++ pyhodl/models/wallets.py
@@ -23,7 +23,7 @@ from datetime import datetime
 
 import numpy as np
 
-from pyhodl.apis.prices import get_price
+from pyhodl.apis.prices.utils import get_price
 from pyhodl.config import VALUE_KEY, DATE_TIME_KEY
 from pyhodl.data.tables import get_coin_prices_table
 from pyhodl.utils import is_crypto, is_nan
